+++ chalicelib/compliance.py
@@ -5,7 +5,7 @@ from . import pvsecret
 from chalicelib.version import API_VERSION
 from chalicelib.telemetry import capture_metric, CostMetrics, InfoMetrics
 from chalicelib.usage import get_openai_usage, get_boost_cost, OpenAIDefaults
-from chalicelib.payments import update_usage_for_code
+from chalicelib.payments import update_usage_for_text
 
 compliance_api_version = API_VERSION  # API version is global for now, not service specific
 print("compliance_api_version: ", compliance_api_version)
@@ -80,7 +80,7 @@ def compliance_code(code, account, context, correlation_id):
 
         try:
             # update the billing usage for this analysis
-            update_usage_for_code(account, prompt + explanation)
+            update_usage_for_text(account, prompt + explanation)
         except Exception:
             exception_info = traceback.format_exc()
             print("UPDATE_USAGE:FAILURE:{}:{}:{}:{}:Error updating ~${} usage: ".format(customer['name'], customer['id'], email, correlation_id, boost_cost), exception_info)