+++ chalicelib/convert.py
@@ -5,7 +5,7 @@ import os
 from chalicelib.version import API_VERSION
 from chalicelib.telemetry import capture_metric, CostMetrics, InfoMetrics
 from chalicelib.usage import get_openai_usage, get_boost_cost, OpenAIDefaults
-from chalicelib.payments import update_usage_for_code
+from chalicelib.payments import update_usage_for_text
 
 
 secret_json = pvsecret.get_secrets()
@@ -97,7 +97,7 @@ def explain_code(code, account, context, correlation_id):
 
         try:
             # update the billing usage for this analysis
-            update_usage_for_code(account, prompt + explanation)
+            update_usage_for_text(account, prompt + explanation)
         except Exception:
             exception_info = traceback.format_exc()
             print("UPDATE_USAGE:FAILURE:{}:{}:{}:{}:Error updating ~${} usage: ".format(customer['name'], customer['id'], email, correlation_id, boost_cost), exception_info)
@@ -176,7 +176,7 @@ def generate_code(summary, original_code, language, account, context, correlatio
 
         try:
             # update the billing usage for this analysis
-            update_usage_for_code(account, prompt + generate_code)
+            update_usage_for_text(account, prompt + generate_code)
         except Exception:
             exception_info = traceback.format_exc()
             print("UPDATE_USAGE:FAILURE:{}:{}:{}:{}:Error updating ~${} usage: ".format(customer['name'], customer['id'], email, correlation_id, boost_cost), exception_info)