+++ chalicelib/customprocess.py
@@ -5,7 +5,7 @@ import os
 from chalicelib.version import API_VERSION
 from chalicelib.telemetry import capture_metric, CostMetrics, InfoMetrics
 from chalicelib.usage import get_openai_usage, get_boost_cost, OpenAIDefaults
-from chalicelib.payments import update_usage_for_code
+from chalicelib.payments import update_usage_for_text
 
 customprocess_api_version = API_VERSION  # API version is global for now, not service specific
 print("customprocess_api_version: ", customprocess_api_version)
@@ -84,7 +84,7 @@ def customprocess_code(code, customprompt, account, context, correlation_id):
 
         try:
             # update the billing usage for this analysis
-            update_usage_for_code(account, prompt + analysis)
+            update_usage_for_text(account, prompt + analysis)
         except Exception:
             exception_info = traceback.format_exc()
             print("UPDATE_USAGE:FAILURE:{}:{}:{}:{}:Error updating ~${} usage: ".format(customer['name'], customer['id'], email, correlation_id, boost_cost), exception_info)