+++ tests/test_payments.py
@@ -2,7 +2,7 @@
 import stripe
 
 # Import the checkCreateCustomer function from the payments module in the chalicelib directory
-from chalicelib.payments import check_create_customer, check_create_subscription, check_create_subscription_item, update_usage, check_trial_expired, check_valid_subscriber
+from chalicelib.payments import check_create_customer, check_create_subscription, check_create_subscription_item, update_usage, check_customer_account_status, check_valid_subscriber
 
 
 # utility function to generate an org name with a random domain
@@ -135,8 +135,9 @@ def test_update_usage_large():
     subscription_item = check_create_subscription_item(subscription=subscription, email=email)
 
     # now check that we correctly flag the customer as in a trial
-    expired = check_trial_expired(customer=customer)
-    assert expired is False
+    active, account_status = check_customer_account_status(customer=customer)
+    assert active is True
+    assert account_status == 'trial'
     # Call the updateUsage function with the test inputs
     result = update_usage(subscription_item=subscription_item, bytes=104857600)
 
@@ -152,8 +153,9 @@ def test_update_usage_large():
     assert customer.balance is not None
 
     # now check that we correctly flag the customer as needing to be charged
-    expired = check_trial_expired(customer=customer)
-    assert expired is True
+    active, account_status = check_customer_account_status(customer=customer)
+    assert active is False
+    assert account_status == 'suspended'
 
     # get the pending invoice
     invoice = stripe.Invoice.upcoming(customer=customer.id)